+++ src/main/pages/Home/LinkList/Header/index.jsx
@@ -9,6 +9,8 @@ import { useToast } from '@modules/ui'
 import { createTabList } from '@utils/chromeApis/tab'
 import { GAEvent } from '@utils/ga'
 
+import AlarmList from './AlarmList'
+import DraggableHistoryList from './DraggableHistoryList'
 import EditableCategoryTitle from './EditableCategoryTitle'
 import useStyles from './style'
